backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Celery and Redis
celery==5.3.6
redis==5.0.1

# Database
pymongo==4.6.1
motor==3.3.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Supabase
supabase==2.3.4

# Gmail API
google-api-python-client==2.111.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# PDF and CSV Processing
PyPDF2==3.0.1
pdfplumber==0.10.4
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1  # For .xls (old Excel format) support

# ML Libraries
scikit-learn==1.4.1.post1
numpy==1.26.3
pandas==2.1.4

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
httpx<0.26.0

# Real-time
websockets==12.0
python-socketio==5.11.0

# Date utilities
python-dateutil==2.8.2

# File type detection
python-magic==0.4.27

# Async utilities
anyio==4.2.0

